+++ src/rana_process_sdk/domain/dataset.py
@@ -12,6 +12,14 @@ class ResourceIdentifier(BaseModel):
     link: AnyHttpUrl | None
 
 
+class DatasetLink(BaseModel):
+    model_config = ConfigDict(frozen=True)
+
+    protocol: str
+    name: str | None = None
+    url: AnyHttpUrl | None = None
+
+
 class RanaDataset(BaseModel):
     """Subset of fields of the Dataset get and search response"""
 
@@ -20,12 +28,27 @@ class RanaDataset(BaseModel):
     id: str
     title: str  # mapped from resourceTitleObject.default
     resource_identifier: list[ResourceIdentifier] = []
+    links: list[DatasetLink] = []
 
     def get_id_for_namespace(self, namespace: AnyHttpUrl) -> str | None:
         """Returns the id of given namespace, otherwise None."""
         for identifier in self.resource_identifier:
             if identifier.link == namespace:
                 return identifier.code
+        return None
+
+    def get_wcs_link(self) -> DatasetLink | None:
+        """Returns the WCS link if available, otherwise None."""
+        for link in self.links:
+            if link.protocol == "OGC:WCS":
+                return link
+        return None
+
+    def get_wfs_link(self) -> DatasetLink | None:
+        """Returns the WFS link if available, otherwise None."""
+        for link in self.links:
+            if link.protocol == "OGC:WFS":
+                return link
         return None