+++ src/rana_process_sdk/settings/local_test_settings.py
@@ -9,6 +9,7 @@ from pydantic_settings import (
 )
 from pydantic_settings_yaml.base_settings import YamlConfigSettingsSource
 
+from ..domain import RanaDataset
 from .settings import LizardSettings
 
 __all__ = ["get_local_test_settings", "LocalTestSettings"]
@@ -20,16 +21,10 @@ class TestThreediSettings(BaseModel):
     organisation: UUID
 
 
-class TestDataset(BaseModel):
-    id: str
-    title: str
-    lizard_raster_id: str
-
-
 class LocalTestSettings(BaseSettings):
-    lizard: LizardSettings
-    threedi: TestThreediSettings
-    datasets: dict[str, TestDataset] = {}
+    lizard: LizardSettings | None = None
+    threedi: TestThreediSettings | None = None
+    datasets: dict[str, RanaDataset] = {}
 
     @classmethod
     def settings_customise_sources(
@@ -63,7 +58,7 @@ class LocalTestSettings(BaseSettings):
 
     model_config = SettingsConfigDict(
         secrets_dir=environ.get("SETTINGS_SECRETS_DIR", "/etc/secrets"),
-        yaml_file="/code/local_test/config.yaml",
+        yaml_file="config.yaml",
         env_prefix="RANA_",
         env_nested_delimiter="__",
     )