+++ tests/domain/test_dataset.py
@@ -1,7 +1,7 @@
 from pydantic import AnyHttpUrl
 from pytest import fixture, mark
 
-from rana_process_sdk.domain.dataset import RanaDataset, ResourceIdentifier
+from rana_process_sdk.domain.dataset import DatasetLink, RanaDataset, ResourceIdentifier
 
 
 @fixture
@@ -12,6 +12,18 @@ def dataset() -> RanaDataset:
         resource_identifier=[
             ResourceIdentifier(code="id-1", link=AnyHttpUrl("https://namespace/1")),
             ResourceIdentifier(code="id-2", link=AnyHttpUrl("https://namespace/2")),
+        ],
+        links=[
+            DatasetLink(
+                protocol="OGC:WCS",
+                name="example",
+                url=AnyHttpUrl("https://example.com/wcs"),
+            ),
+            DatasetLink(
+                protocol="OGC:WFS",
+                name="other",
+                url=AnyHttpUrl("https://example.com/wfs"),
+            ),
         ],
     )
 
@@ -28,3 +40,21 @@ def test_get_id_for_namespace(
     dataset: RanaDataset, namespace: AnyHttpUrl, expected_id: str | None
 ) -> None:
     assert dataset.get_id_for_namespace(namespace) == expected_id
+
+
+def test_get_wcs_link(dataset: RanaDataset) -> None:
+    assert dataset.get_wcs_link() == dataset.links[0]
+
+
+def test_get_wfs_link(dataset: RanaDataset) -> None:
+    assert dataset.get_wfs_link() == dataset.links[1]
+
+
+def test_get_wcs_link_no_links(dataset: RanaDataset) -> None:
+    del dataset.links[0]
+    assert dataset.get_wcs_link() is None
+
+
+def test_get_wfs_link_no_links(dataset: RanaDataset) -> None:
+    del dataset.links[1]
+    assert dataset.get_wfs_link() is None