+++ CHANGES.md
@@ -4,7 +4,11 @@
 ## 0.2 (unreleased)
 
 
-- Nothing changed yet.
+- Added `context.get_dataset`.
+
+- Add support for retrieving WCS and WFS links for Rana datasets.
+
+- Change the test settings model.
 
 
 ## 0.1 (2025-09-16)